feat: Extract function to format card byline
