+++ blocks/stories/stories.js
@@ -2,6 +2,7 @@ import {
   createOptimizedPicture, lookupPages,
 } from '../../scripts/scripts.js';
 import colormap from '../../scripts/colormap.js';
+import formatCardByline from '../../scripts/format-card-byline.js';
 import tagLink from '../../scripts/tag-link.js';
 
 function createCard(row) {
@@ -13,17 +14,7 @@ function createCard(row) {
   imageLink.append(createOptimizedPicture(row.image, row.title));
 
   const cardSubTitle = `${row.subtitle ? `<p class="cmp-stories-card__intro">${row.subtitle}</p>` : ''}`;
-
-  const authors = row.author.split(', ');
-  const authorTitles = row.authorTitle.split('; ');
-  let cardByline = '';
-  authors.forEach((author, idx) => {
-    cardByline += `
-      ${author ? `<p class="cmp-stories-card__author">${author}</p>` : ''}
-      ${authorTitles[idx] ? `<p class="cmp-stories-card__author-title">${authorTitles[idx]}</p>` : ''}
-    `;
-  });
-
+  const cardByline = formatCardByline(row);
   const cardBGColor = row.color !== '' ? row.color : '#fff';
   const textColor = colormap[cardBGColor];
   const cardTag = row.tag !== '' ? `${row.tag}` : '';
